Se cierra pedido final, con logica funcionando ok.

+++ Dominio/PedidosDetalle.cs
@@ -12,7 +12,7 @@ namespace Dominio
         public int PedidoID { get; set; }
         public int ProductoID { get; set; }
         public int Cantidad { get; set; }
-        public float PrecioUnitario { get; set; }
+        public decimal PrecioUnitario { get; set; }
         public Productos producto { get; set; }
     }
 }
